+++ lib/grape/dsl/routing.rb
@@ -82,7 +82,7 @@ module Grape
             in_setting = inheritable_setting
 
             if app.respond_to?(:inheritable_setting, true)
-              mount_path = Rack::Mount::Utils.normalize_path(path)
+              mount_path = Grape::Router.normalize_path(path)
               app.top_level_setting.namespace_stackable[:mount_path] = mount_path
 
               app.inherit_settings(inheritable_setting)
@@ -98,6 +98,7 @@ module Grape
               method: :any,
               path: path,
               app: app,
+              forward_match: !app.respond_to?(:inheritable_setting),
               for: self
             )
           end