+++ lib/grape/middleware/versioner/accept_version_header.rb
@@ -14,7 +14,7 @@ module Grape
       #    env['api.version']  => 'v1'
       #
       # If version does not match this route, then a 406 is raised with
-      # X-Cascade header to alert Rack::Mount to attempt the next matched
+      # X-Cascade header to alert Grape::Router to attempt the next matched
       # route.
       class AcceptVersionHeader < Base
         def before
@@ -46,7 +46,7 @@ module Grape
         end
 
         # By default those errors contain an `X-Cascade` header set to `pass`, which allows nesting and stacking
-        # of routes (see [Rack::Mount](https://github.com/josh/rack-mount) for more information). To prevent
+        # of routes (see Grape::Router) for more information). To prevent
         # this behavior, and not add the `X-Cascade` header, one can set the `:cascade` option to `false`.
         def cascade?
           if options[:version_options] && options[:version_options].key?(:cascade)